% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrewrt.R
\name{lm.rewrt}
\alias{lm.rewrt}
\title{Fitting Linear Model}
\usage{
lm.rewrt(formula, data, prt = FALSE)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted with specific pattern (i.e. y ~ x1 + x2 where y is the responding variable; x1 and x2 are the covariates for the model)}

\item{data}{an optional data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from formula.}

\item{prt}{logical; if TRUE, a formated output will be printed on the screen.}
}
\value{
A list containing the following elements
        \itemize{
           \item call - return the fitted linear model fomula and the corresponding data.
           \item coefficients - a table with the estimated values for each covariates and the intercept as well as their corresponding standard error, t-statistics and (two-sided) p-value.
           \item residuals - summary of the usual residuals with min, 1st quantile, median, 3rd quantile, max being computed.
           \item fitted.values - the fitted mean values.
           \item design.matrix - the design matrix for model fitting
           \item Residual standard erro - the square root of the estimated variance of the random error and a corresponding degrees of freedom will also be computed.
           \item r.squared - R^2, the 'fraction of variance explained by the model', SSR (variation in fitted values about the overall mean) / SSY (total variation in Y about its overall mean).
           \item adj.r.squared - adjusted version of R^2, penalized based on number of covariates.
           \item cov.unscaled - a table of (unscaled) covariances of the coeficients.
           \item sd.beta - the corresponding standard error for the estimated coefficient.
           \item t value - the t-statistic for corresponding variable.
           \item Pr(>|t|) - the corresponding (two-sided) p-value for the t-statistic
           \item fstatistic - the test statistic for F-tests. Variation Between Sample Means / Variation Within the Samples.
        }
}
\description{
"lm.rewrt" is used to fit simple linear model.
}
