% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrewrt.R
\name{summary.lm.rewrt}
\alias{summary.lm.rewrt}
\title{Summarizing Linear Model Fits}
\usage{
\method{summary}{lm.rewrt}(object, correlation = FALSE, prt = TRUE)
}
\arguments{
\item{object}{the fitting results of "lm.rewrt".}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.}

\item{prt}{logical; if TRUE, a summarized table will be computed to the screen}
}
\value{
A list containing the following elements
        \itemize{
           \item call - the fitted linear model fomula.
           \item Residuals - summary of the usual residuals with min, 1st quantile, median, 3rd quantile, max being computed.
           \item Coefficients - a table with the estimated values for each covariates and the intercept as well as their corresponding standard error, t-statistics and (two-sided) p-value.
           \item Residual standard erro - the square root of the estimated variance of the random error and a corresponding degrees of freedom will also be computed.
           \item r.squared - R^2, the 'fraction of variance explained by the model', SSR (variation in fitted values about the overall mean) / SSY (total variation in Y about its overall mean).
           \item adj.r.squared - adjusted version of R^2, penalized based on number of covariates.
           \item cov.unscaled - a table of (unscaled) covariances of the coeficients.
           \item correlation - the correlation table corresponding to the above cov.unscaled table, if correlation = TRUE is specified.
        }
}
\description{
Summarize the output of "lm.rewrt" function
}
